rename `circuitcapacitychecker` to `ccc`
