+++ .github/workflows/ci.yml
@@ -46,8 +46,8 @@ jobs:
     - name: Build
       run: |
         make libzkp
-        sudo cp ./rollup/circuitcapacitychecker/libzkp/libzkp.so /usr/local/lib/
-        sudo cp ./rollup/circuitcapacitychecker/libzkp/libzktrie.so /usr/local/lib/
+        sudo cp ./rollup/ccc/libzkp/libzkp.so /usr/local/lib/
+        sudo cp ./rollup/ccc/libzkp/libzktrie.so /usr/local/lib/
         make geth
   # check:
   #   if: github.event.pull_request.draft == false
