+++ Dockerfile
@@ -8,15 +8,15 @@ FROM scrolltech/go-rust-builder:go-1.21-rust-nightly-2023-12-03 as chef
 WORKDIR app
 
 FROM chef as planner
-COPY ./rollup/circuitcapacitychecker/libzkp/ .
+COPY ./rollup/ccc/libzkp/ .
 RUN cargo chef prepare --recipe-path recipe.json
 
 FROM chef as zkp-builder
-COPY ./rollup/circuitcapacitychecker/libzkp/rust-toolchain ./
+COPY ./rollup/ccc/libzkp/rust-toolchain ./
 COPY --from=planner /app/recipe.json recipe.json
 RUN cargo chef cook --release --recipe-path recipe.json
 
-COPY ./rollup/circuitcapacitychecker/libzkp .
+COPY ./rollup/ccc/libzkp .
 RUN cargo clean
 RUN cargo build --release
